requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-PyMongo==3.0.1
pymongo==4.15.3
opencv-python-headless==4.12.0.88
numpy==2.2.6
Pillow==11.1.0
python-dotenv==1.2.1
dnspython==2.8.0
click==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
gunicorn==23.0.0
bcrypt==4.1.2
requests==2.31.0
